Renamed configuration explicitReturnTypes to explicitResultTypes
