+++ scalafix-tests/input/src/main/scala/test/explicitResultTypes/ExplicitResultTypesShort.scala
@@ -1,8 +1,8 @@
 /*
 rules = ExplicitResultTypes
-explicitReturnTypes.unsafeShortenNames = true
+explicitResultTypes.unsafeShortenNames = true
  */
-package test.explicitReturnTypes
+package test.explicitResultTypes
 
 import scala.language.implicitConversions
